Restore config to match changes prior to v11.0.7
Removed DNS changes under vendor/overlay as they are incomplete compared
to 00004-use-cloudflare-dns.patch. Will have to investigate whether it's
possible to override the .java files like you're able to with the .xml
files at a later date.

Have left the addition of SeedVault as it's implementation appears to be
complete and in which I consider an essential application.